#150 abstract creation of fetch GET options for sql api calls
