+++ app/src/utils/cartoSqlApiAuth.js
@@ -1,0 +1,18 @@
+import { cartoApiKey } from "../constants/config";
+
+/**
+ * creates the expected fetch request options for Carto SQL API calls
+ * @returns {object}
+ */
+export const cartoSqlApiAuthOptions = () => {
+  const headers = new Headers();
+  headers.append("Authorization", `Bearer ${cartoApiKey}`);
+
+  const requestOptions = {
+    method: "GET",
+    headers: headers,
+    redirect: "follow",
+  };
+
+  return requestOptions;
+};
